Contentful Integration
Needs x-browser support (es6 code failing in IE)
Removed nav bar and footer
JS could be further optimized and minified
Note partner image links are pulled from the "description" field of
that image asset
NOTE: meta data is not currently part of the contentful content model,
nor is it dynamically updated when the entry is pulled down. This is
definitely a high-priority to-do.